src/requirements.txt
argh==0.26.2
certifi==2019.6.16
chardet==3.0.4
h5json==1.1.3
h5py==2.10.0
h5pyd==0.4.4
idna==2.8
mock==3.0.5
numexpr==2.6.9
numpy==1.17.0
pathtools==0.1.2
Pillow==6.1.0
PyQt5==5.13.0
PyQt5-sip==4.19.18
pytz==2019.2
PyYAML==5.1.2
QtPy==1.9.0
requests==2.22.0
six==1.12.0
tables==3.5.2
tornado==6.0.3
urllib3==1.25.3
ViTables==3.0.0
watchdog==0.9.0
